Added JPA staff, and Cancha entity, Controller and Repository
