+++ src/main/java/com/personal/api/usermanager/UserManagerApplication.java
@@ -6,7 +6,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
 @SpringBootApplication
-@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
 public class UserManagerApplication {
 
 	public static void main(String[] args) {
