+++ src/main/java/com/personal/api/usermanager/model/User.java
@@ -1,0 +1,35 @@
+package com.personal.api.usermanager.model;
+
+public class User {
+	
+	private int id;
+
+    private String firstName;
+
+    private String lastName;
+    
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+    
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+}
