+++ src/main/resources/application.properties
@@ -1,1 +1,7 @@
 server.port=5000
+#spring.datasource.url: jdbc:mysql://${MYSQL_HOST:localhost}:${MYSQL_PORT:3306}/${MYSQL_DB:usermanager}
+#spring.datasource.username:${MYSQL_USERNAME:root}
+#spring.datasource.password:${MYSQL_PASSWORD:root}
+spring.datasource.url: jdbc:mysql://aa19efn0m9usu05.c28q6sdvw4oc.us-east-2.rds.amazonaws.com:3306/usermanager
+spring.datasource.username:usermanager
+spring.datasource.password:Welcome1!
