+++ src/test/java/com/personal/api/usermanager/UserManagerApplicationTests.java
@@ -5,12 +5,12 @@ import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class UserManagerApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
+//@RunWith(SpringRunner.class)
+//@SpringBootTest
+//public class UserManagerApplicationTests {
+//
+////	@Test
+////	public void contextLoads() {
+////	}
+//
+//}
